+++ streaming_feature_simulator/sql_template/checkpoint_rekarma.sql
@@ -1,6 +1,6 @@
 WITH tmp_source AS (
     SELECT par_process_date
-         , key_event_info_event_time AS checkpoint_time
+         , key_event_info_event_time::BIGINT AS checkpoint_time
          , key_consumer_uuid
          , key_order_token
          , feature_map
